feat: Use bitmaps for boot logo replacement (#1370)

+++ src/boot.sh
@@ -133,29 +133,28 @@ prepareUefiRom() {
 
   [ -s "$DEST.rom" ] && return 0
 
-  [ ! -s "$OVMF/$ROM" ] && error "UEFI boot file ($OVMF/$ROM) not found!" && exit 44
+  local rom="$OVMF/$ROM"
+  [ ! -s "$rom" ] && error "UEFI boot file ($rom) not found!" && exit 44
 
-  local logo="/var/www/img/${PROCESS,,}.ffs"
-  [ ! -s "$logo" ] && logo="/var/www/img/qemu.ffs"
-  [ ! -s "$logo" ] && LOGO="N"
+  local logo="/var/www/img/${PROCESS,,}.bmp"
+  [ ! -s "$logo" ] && logo="/var/www/img/qemu.bmp"
+
+  if ! disabled "$LOGO" && [ ! -s "$logo" ]; then
+    LOGO="N"
+    warn "boot logo file ($logo) not found!"
+  fi
 
   rm -f "$DEST.tmp"
 
-  if disabled "$LOGO"; then
-    if ! cp "$OVMF/$ROM" "$DEST.tmp"; then
-      rm -f "$DEST.tmp"
-      error "Failed to copy UEFI boot file to $DEST.tmp" && exit 44
-    fi
-  else
-    if ! /run/utk.bin "$OVMF/$ROM" replace_ffs LogoDXE "$logo" save "$DEST.tmp"; then
-      warn "failed to add custom logo to BIOS!"
-      rm -f "$DEST.tmp"
+  if ! disabled "$LOGO" &&
+     ! /run/boot-logo "$logo" "$rom" --output "$DEST.tmp" -q; then
+    warn "failed to add custom logo ($logo) to BIOS!"
+    rm -f "$DEST.tmp"
+  fi
 
-      if ! cp "$OVMF/$ROM" "$DEST.tmp"; then
-        rm -f "$DEST.tmp"
-        error "Failed to copy UEFI boot file to $DEST.tmp" && exit 44
-      fi
-    fi
+  if [[ ! -f "$DEST.tmp" ]] && ! cp "$rom" "$DEST.tmp"; then
+    rm -f "$DEST.tmp"
+    error "Failed to copy UEFI boot file to $DEST.tmp" && exit 44
   fi
 
   if ! mv "$DEST.tmp" "$DEST.rom"; then
@@ -177,11 +176,12 @@ prepareUefiVars() {
 
   [ -s "$DEST.vars" ] && return 0
 
-  [ ! -s "$OVMF/$VARS" ] && error "UEFI vars file ($OVMF/$VARS) not found!" && exit 45
+  local vars="$OVMF/$VARS"
+  [ ! -s "$vars" ] && error "UEFI vars file ($vars) not found!" && exit 45
 
   rm -f "$DEST.tmp"
 
-  if ! cp "$OVMF/$VARS" "$DEST.tmp"; then
+  if ! cp "$vars" "$DEST.tmp"; then
     rm -f "$DEST.tmp"
     error "Failed to copy UEFI vars file to $DEST.tmp" && exit 45
   fi
